requirements.txt
google-api-core
google-auth
google-cloud-core
google-cloud-documentai
google-cloud-documentai-toolbox
google-cloud-storage
google-cloud-vision
google-crc32c
google-resumable-media
googleapis-common-protos
grpc-google-iam-v1
grpcio
grpcio-status
requests
